+++ src/game/g_script.c
@@ -1279,7 +1279,7 @@ void SP_script_mover(gentity_t *ent)
 		if (G_SpawnString("description", "", &s))
 		{
 			trap_GetConfigstring(CS_SCRIPT_MOVER_NAMES, cs, sizeof(cs));
-			Info_SetValueForKey(cs, va("%i", ent - g_entities), s);
+			Info_SetValueForKey(cs, va("%i", (int)(ent - g_entities)), s);
 			trap_SetConfigstring(CS_SCRIPT_MOVER_NAMES, cs);
 		}
 	}
